Web: Transactions page improved, Transaction Filter added to search by any property (Description, Location, Dates, Amounts). API: Methods added to retrieve summary by properties (categories, payment methods) and also total values by specific properties (categoryId, paymentMethodId)

+++ BaseLibrary/Entities/Category.cs
@@ -7,8 +7,8 @@ namespace BaseLibrary.Entities
     public class Category
     {
         [Key]        
-        public int Id { get; set; }        
-        public int? TransactionTypeId { get; set; }
+        public byte Id { get; set; }        
+        public byte? TransactionTypeId { get; set; }
         [Column(TypeName = "nvarchar(50)")]
         public string Name { get; set; }
